+++ ui/src/main.tsx
@@ -10,8 +10,8 @@ import '@fontsource-variable/inter'
 import './assets/icon.png'
 import './index.css'
 
+// Hydrate store from server data
 {
-  // Hydrate store from server data
   const element = document.getElementById('__FROG_DATA__')
   const bootstrap = JSON.parse(element!.textContent!) as Bootstrap
   hydrateStore(bootstrap)
@@ -27,3 +27,9 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
 
 // Watch for changes in development
 initFrogClient()
+
+// Remove server style
+{
+  const element = document.getElementById('__SSR_STYLE__')
+  element?.remove()
+}